id: split into modules
The main file was getting complicated, so split it into different files.
